refactor(desktop): hoist duration formatters into lib/format-duration
Move formatDurationMs / formatRunningDurationMs out of EnvActionRunsView
into a shared renderer lib module so the pricing panel can reuse the same
2h 3m 4s duration vocabulary without a feature-to-feature import.
EnvActionRunsView re-exports them so Composer and its tests are untouched.
