+++ apps/desktop/src/renderer/src/features/thread-detail/EnvActionRunsView.tsx
@@ -8,6 +8,14 @@ import {
 } from "react";
 import type { CodexEnvironmentActionRun } from "@pwragent/shared";
 import { useViewportTooltip } from "../../lib/useViewportTooltip";
+import {
+  formatDurationMs,
+  formatRunningDurationMs,
+} from "../../lib/format-duration";
+
+// Re-exported so existing importers (Composer + its tests) keep resolving the
+// duration helpers from here; the source of truth now lives in lib/format-duration.
+export { formatDurationMs, formatRunningDurationMs };
 
 const ENV_ACTION_OUTPUT_MAX_LINES = 500;
 
@@ -79,37 +87,6 @@ function elementContainsSelection(element: HTMLElement | null): boolean {
     }
   }
   return false;
-}
-
-export function formatDurationMs(
-  ms?: number,
-  options?: { coarseAfterMinute?: boolean },
-): string {
-  if (!ms || !Number.isFinite(ms)) return "";
-  if (ms < 1_000) return `${Math.round(ms)}ms`;
-  const totalSeconds = Math.round(ms / 1_000);
-  if (totalSeconds < 60) return `${totalSeconds}s`;
-  const totalMinutes = Math.floor(totalSeconds / 60);
-  const seconds = totalSeconds % 60;
-  if (totalMinutes < 60) {
-    if (options?.coarseAfterMinute) return `${totalMinutes}m`;
-    return seconds ? `${totalMinutes}m ${seconds}s` : `${totalMinutes}m`;
-  }
-  const totalHours = Math.floor(totalMinutes / 60);
-  const minutes = totalMinutes % 60;
-  if (totalHours < 24) {
-    return minutes ? `${totalHours}h ${minutes}m` : `${totalHours}h`;
-  }
-  const days = Math.floor(totalHours / 24);
-  const hours = totalHours % 24;
-  return hours ? `${days}d ${hours}h` : `${days}d`;
-}
-
-export function formatRunningDurationMs(ms?: number): string {
-  if (ms === undefined || !Number.isFinite(ms) || ms < 0) return "";
-  const totalSeconds = Math.floor(ms / 1_000);
-  if (totalSeconds < 1) return "0s";
-  return formatDurationMs(totalSeconds * 1_000);
 }
 
 export function EnvActionRunsView(props: {
